objtool: orc: Refactor ORC API for other architectures to implement.
The ORC unwinder is only supported on x86 at the moment and should thus be
in the x86 architecture code. In order not to break the whole structure in
case another architecture decides to support the ORC unwinder via objtool,
move the implementation be done in the architecture dependent code.

Signed-off-by: Raphael Gault <raphael.gault@arm.com>
[J.T. Rebase on orc name changes,
      Move arch_orc_read_unwind_hints() to orc.h,
      Reword commit message to use imperative language]
Signed-off-by: Julien Thierry <jthierry@redhat.com>