objtool: check: Validate local symbols getting called
Non-function local symbols do not get validate if they are only
referenced by a call. Currently, adding an unwind_hint is necessary to
have them validated.

It would be incorrect to validate callable exported symbol not
conforming to the function standard without an unwind hint. But it
is valid to analyze local symbols using the state(s) at their call
site(s).

This reduces the amount of unwind hints needed.

Signed-off-by: Julien Thierry <jthierry@redhat.com>